Time: 31 ms (5.33%), Space: 22.9 MB (45.76%) - LeetHub

+++ 0875-koko-eating-bananas/0875-koko-eating-bananas.cpp
@@ -5,7 +5,7 @@ public:
         int right  = *max_element(piles.begin(),piles.end());
 
         while(left < right){
-            int mid = left + (right - left)/2;
+            long mid = left + (right - left)/2;
             long long hours = 0;
 
             for(auto banana : piles){
